Consolidate UPath, test fsspec local
