+++ upath/__init__.py
@@ -2,3 +2,5 @@
 __version__ = "0.0.11"
 
 from upath.core import UPath
+
+__all__ = ["UPath"]
