+++ upath/implementations/hdfs.py
@@ -1,7 +1,7 @@
-from upath.universal_path import _FSSpecAccessor, UniversalPath
+import upath.core
 
 
-class _HDFSAccessor(_FSSpecAccessor):
+class _HDFSAccessor(upath.core._FSSpecAccessor):
     def __init__(self, parsed_url, *args, **kwargs):
         super().__init__(parsed_url, *args, **kwargs)
         self._fs.root_marker = "/"
@@ -11,7 +11,6 @@ class _HDFSAccessor(_FSSpecAccessor):
         argument is a UniversalPath instance, that argument is replaced with
         the UniversalPath's path attribute
         """
-
         def wrapper(*args, **kwargs):
             args, kwargs = self._transform_arg_paths(args, kwargs)
             if "trunicate" in kwargs:
@@ -23,5 +22,5 @@ class _HDFSAccessor(_FSSpecAccessor):
         return wrapper
 
 
-class HDFSPath(UniversalPath):
+class HDFSPath(upath.core.UPath):
     _default_accessor = _HDFSAccessor
