+++ upath/implementations/http.py
@@ -1,9 +1,9 @@
 import urllib
 
-from upath.universal_path import UniversalPath, _FSSpecAccessor
+import upath.core
 
 
-class _HTTPAccessor(_FSSpecAccessor):
+class _HTTPAccessor(upath.core._FSSpecAccessor):
     def __init__(self, parsed_url, *args, **kwargs):
         super().__init__(parsed_url, *args, **kwargs)
 
@@ -12,13 +12,12 @@ class _HTTPAccessor(_FSSpecAccessor):
         argument is a UniversalPath instance, that argument is replaced with
         the UniversalPath's path attribute
         """
-
         def wrapper(*args, **kwargs):
             if args:
                 args = list(args)
                 first_arg = args.pop(0)
                 if not kwargs.get("path"):
-                    if isinstance(first_arg, UniversalPath):
+                    if isinstance(first_arg, upath.core.UPath):
                         first_arg = str(first_arg)
                         args.insert(0, first_arg)
                     args = tuple(args)
@@ -31,5 +30,5 @@ class _HTTPAccessor(_FSSpecAccessor):
         return wrapper
 
 
-class HTTPPath(UniversalPath):
+class HTTPPath(upath.core.UPath):
     _default_accessor = _HTTPAccessor
